requirements.txt
python-nmap
prettyTables